Minimal example produces no graphics on Linux
This is a minimal example which shows different behavior when run on
Linux vs Windows. Running on Windows produces 2 images with blue
rectangles (out1.png and out2.png) contrary to Linux where out1.png is
empty. out2.png is a blue rectangle on Linux if the bitmap get saved to
disk before it is cloned (the copy saved to disk is removed before the
clone operation).